moved test language model to config so it is available to all tests
